+++ src/analyze/Resolution.hpp
