+++ src/analyze/ResourceManager.hpp
