+++ src/base/Axis.cpp
@@ -1,3 +1,3 @@
-#include "Axis.h"
+#include "Axis.hpp"
 
 ClassImp(Qn::Analysis::Base::AxisConfig)
