+++ src/base/Cut.cpp
@@ -1,3 +1,3 @@
-#include "Cut.h"
+#include "Cut.hpp"
 
 ClassImp(Qn::Analysis::Base::CutConfig)
