+++ src/base/Histogram.cpp
@@ -2,4 +2,4 @@
 // Created by eugene on 12/08/2020.
 //
 
-#include "Histogram.h"
+#include "Histogram.hpp"
