+++ src/base/Variable.hpp
