+++ src/config/Convert.hpp
@@ -5,11 +5,11 @@
 #ifndef FLOW_SRC_CONFIG_CONVERT_H
 #define FLOW_SRC_CONFIG_CONVERT_H
 
-#include <AnalysisSetup.h>
+#include <AnalysisSetup.hpp>
 #include <AnalysisTree/Variable.hpp>
-#include <Histogram.h>
-#include <base/QVector.h>
-#include <base/Variable.h>
+#include <Histogram.hpp>
+#include <base/QVector.hpp>
+#include <base/Variable.hpp>
 #include <bitset>
 
 #include <QnTools/CorrectionOnQnVector.hpp>
