+++ src/correct/CMakeLists.txt
@@ -3,14 +3,14 @@ set(SOURCES
         Task.cpp
         )
 
-string(REPLACE ".cpp" ".h" HEADERS "${SOURCES}")
+string(REPLACE ".cpp" ".hpp" HEADERS "${SOURCES}")
 
 add_library(QnAnalysisCorrect SHARED ${SOURCES} G__QnAnalysisCorrect.cxx)
 link_directories(${PROJECT_LINK_DIRECTORIES})
 add_dependencies(QnAnalysisCorrect ${PROJECT_DEPENDENCIES} QnAnalysisBase)
 include_directories(${PROJECT_INCLUDE_DIRECTORIES} ${CMAKE_CURRENT_SOURCE_DIR})
 
-ROOT_GENERATE_DICTIONARY(G__QnAnalysisCorrect ${HEADERS} LINKDEF QnAnalysisCorrectLinkDef.h)
+ROOT_GENERATE_DICTIONARY(G__QnAnalysisCorrect ${HEADERS} LINKDEF QnAnalysisCorrectLinkDef.hpp)
 target_link_libraries(QnAnalysisCorrect QnAnalysisBase QnAnalysisConfig AnalysisTreeBase AnalysisTreeInfra ${ROOT_LIBRARIES} )
 
 install(TARGETS QnAnalysisCorrect EXPORT QnAnalysisCorrectTargets
